+++ feature/settings/src/androidTest/kotlin/org/meshtastic/feature/settings/debugging/DebugFiltersTest.kt
@@ -15,11 +15,13 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.geeksville.mesh.compose
+package org.meshtastic.feature.settings.debugging
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.test.assertIsDisplayed
@@ -35,11 +37,7 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.meshtastic.core.strings.R
-import org.meshtastic.feature.settings.debugging.DebugActiveFilters
-import org.meshtastic.feature.settings.debugging.DebugCustomFilterInput
-import org.meshtastic.feature.settings.debugging.DebugFilterBar
-import org.meshtastic.feature.settings.debugging.DebugViewModel
-import org.meshtastic.feature.settings.debugging.FilterMode
+import org.meshtastic.feature.settings.debugging.DebugViewModel.UiMeshLog
 
 @RunWith(AndroidJUnit4::class)
 class DebugFiltersTest {
@@ -51,20 +49,19 @@ class DebugFiltersTest {
         val context = InstrumentationRegistry.getInstrumentation().targetContext
         val filterLabel = context.getString(R.string.debug_filters)
         composeTestRule.setContent {
-            var filterTexts by
-                androidx.compose.runtime.remember { androidx.compose.runtime.mutableStateOf(listOf<String>()) }
-            var customFilterText by androidx.compose.runtime.remember { androidx.compose.runtime.mutableStateOf("") }
+            var filterTexts by remember { mutableStateOf(listOf<String>()) }
+            var customFilterText by remember { mutableStateOf("") }
             val presetFilters = listOf("Error", "Warning", "Info")
             val logs =
                 listOf(
-                    com.geeksville.mesh.model.DebugViewModel.UiMeshLog(
+                    UiMeshLog(
                         uuid = "1",
                         messageType = "Info",
                         formattedReceivedDate = "2024-01-01 12:00:00",
                         logMessage = "Sample log message",
                     ),
                 )
-            com.geeksville.mesh.ui.debug.DebugFilterBar(
+            DebugFilterBar(
                 filterTexts = filterTexts,
                 onFilterTextsChange = { filterTexts = it },
                 customFilterText = customFilterText,
@@ -82,17 +79,16 @@ class DebugFiltersTest {
         val context = InstrumentationRegistry.getInstrumentation().targetContext
         val activeFiltersLabel = context.getString(R.string.debug_active_filters)
         composeTestRule.setContent {
-            var filterTexts by
-                androidx.compose.runtime.remember { androidx.compose.runtime.mutableStateOf(listOf<String>()) }
-            var customFilterText by androidx.compose.runtime.remember { androidx.compose.runtime.mutableStateOf("") }
+            var filterTexts by remember { mutableStateOf(listOf<String>()) }
+            var customFilterText by remember { mutableStateOf("") }
             Column(modifier = Modifier.padding(16.dp)) {
-                com.geeksville.mesh.ui.debug.DebugActiveFilters(
+                DebugActiveFilters(
                     filterTexts = filterTexts,
                     onFilterTextsChange = { filterTexts = it },
                     filterMode = FilterMode.OR,
                     onFilterModeChange = {},
                 )
-                com.geeksville.mesh.ui.debug.DebugCustomFilterInput(
+                DebugCustomFilterInput(
                     customFilterText = customFilterText,
                     onCustomFilterTextChange = { customFilterText = it },
                     filterTexts = filterTexts,
@@ -115,9 +111,8 @@ class DebugFiltersTest {
         val context = InstrumentationRegistry.getInstrumentation().targetContext
         val activeFiltersLabel = context.getString(R.string.debug_active_filters)
         composeTestRule.setContent {
-            var filterTexts by
-                androidx.compose.runtime.remember { androidx.compose.runtime.mutableStateOf(listOf("A", "B")) }
-            com.geeksville.mesh.ui.debug.DebugActiveFilters(
+            var filterTexts by remember { mutableStateOf(listOf("A", "B")) }
+            DebugActiveFilters(
                 filterTexts = filterTexts,
                 onFilterTextsChange = { filterTexts = it },
                 filterMode = FilterMode.OR,
